v2 of cdem ha integration
