+++ home-assistant/homeassistant/config/custom_components/cdem/const.py
@@ -1,4 +1,10 @@
 """Constant values for CDEM."""
 
 DOMAIN = "cdem"
-TOPIC_PREFIX = "smartmeter/external2"
+TOPIC_PREFIX = "smartmeter/external3"
+
+TOPICS = [
+    "payload",
+#    "stats",
+#    "announce",
+]
